+++ src/GlobalNavContext.js
@@ -1,15 +1,28 @@
-import React, { useState, createContext } from "react";
+import React, { useEffect, useState, createContext } from "react";
 import PropTypes from "prop-types";
 
 export const GlobalNavContext = createContext();
 
 export const GlobalNavContextProvider = function ({ children }) {
 	const [idOfOpenDropdown, setidOfOpenDropdown] = useState(null);
+	const [accountMenuIsExpanded, setAccountMenuIsExpanded] = useState(false);
 
 	const value = {
 		idOfOpenDropdown,
 		setidOfOpenDropdown,
+		accountMenuIsExpanded,
+		setAccountMenuIsExpanded,
 	};
+
+	useEffect(() => {
+		// we know both are now open
+		if (accountMenuIsExpanded && idOfOpenDropdown) {
+			setAccountMenuIsExpanded((prevState) => {
+				// if prevstate is not null, you know it was already open, therefore close it
+				if (prevState) return false;
+			});
+		}
+	}, [accountMenuIsExpanded, idOfOpenDropdown]);
 
 	return (
 		<GlobalNavContext.Provider value={value}>
