+++ src/Header/Nav/Nav.jsx
@@ -19,7 +19,8 @@ function Nav(props) {
 	const context = useContext(GlobalNavContext);
 	const { ref } = useClickOutside(
 		context.idOfOpenDropdown,
-		context.setidOfOpenDropdown
+		context.setidOfOpenDropdown,
+		context.setAccountMenuIsExpanded
 	);
 
 	function handleNavItemClick(e) {
